+++ src/app-layer-parser.c
@@ -53,7 +53,7 @@
 #include "app-layer-nfs-udp.h"
 #include "app-layer-ntp.h"
 #include "app-layer-tftp.h"
-#include "app-layer-ike.h"
+#include "app-layer/ike/parser.h"
 #include "app-layer-krb5.h"
 #include "app-layer-sip.h"
 #include "app-layer-rfb.h"
