+++ src/app-layer/ike/detect-chosen-sa.c
@@ -26,7 +26,7 @@
 #include "detect-parse.h"
 #include "detect-engine.h"
 #include "detect-engine-content-inspection.h"
-#include "detect-ike-chosen-sa.h"
+#include "app-layer/ike/detect-chosen-sa.h"
 #include "app-layer-parser.h"
 #include "util-byte.h"
 #include "util-unittest.h"
