+++ src/app-layer/ike/detect-exch-type.c
@@ -26,7 +26,7 @@
 #include "detect-parse.h"
 #include "detect-engine.h"
 #include "detect-engine-content-inspection.h"
-#include "detect-ike-exch-type.h"
+#include "app-layer/ike/detect-exch-type.h"
 #include "app-layer-parser.h"
 #include "util-byte.h"
 #include "detect-engine-uint.h"
