+++ src/app-layer/ike/detect-key-exchange-payload.c
@@ -43,11 +43,11 @@
 #include "app-layer.h"
 #include "app-layer-parser.h"
 
-#include "detect-ike-key-exchange-payload.h"
+#include "app-layer/ike/detect-key-exchange-payload.h"
 #include "stream-tcp.h"
 
 #include "rust.h"
-#include "app-layer-ike.h"
+#include "app-layer/ike/parser.h"
 #include "rust-bindings.h"
 
 #define KEYWORD_NAME_KEY_EXCHANGE "ike.key_exchange_payload"
