+++ src/detect-engine-register.c
@@ -273,14 +273,14 @@
 #include "detect-modbus.h"
 #include "detect-cipservice.h"
 #include "app-layer/dnp3/detect-dnp3.h"
-#include "detect-ike-exch-type.h"
-#include "detect-ike-spi.h"
-#include "detect-ike-vendor.h"
-#include "detect-ike-chosen-sa.h"
-#include "detect-ike-key-exchange-payload-length.h"
-#include "detect-ike-nonce-payload-length.h"
-#include "detect-ike-nonce-payload.h"
-#include "detect-ike-key-exchange-payload.h"
+#include "app-layer/ike/detect-exch-type.h"
+#include "app-layer/ike/detect-spi.h"
+#include "app-layer/ike/detect-vendor.h"
+#include "app-layer/ike/detect-chosen-sa.h"
+#include "app-layer/ike/detect-key-exchange-payload-length.h"
+#include "app-layer/ike/detect-nonce-payload-length.h"
+#include "app-layer/ike/detect-nonce-payload.h"
+#include "app-layer/ike/detect-key-exchange-payload.h"
 
 #include "action-globals.h"
 #include "tm-threads.h"
