+++ src/rust-context.h
@@ -23,7 +23,7 @@
 #include "detect-engine-state.h" //DetectEngineState
 
 #include "app-layer-krb5.h" //KRB5State, KRB5Transaction
-#include "app-layer-ike.h"  //IKEState, IKETransaction
+#include "app-layer/ike/parser.h" //IKEState, IKETransaction
 #include "app-layer-ntp.h" //NTPState, NTPTransaction
 #include "app-layer-snmp.h" //SNMPState, SNMPTransaction
 #include "app-layer-tftp.h" //TFTPState, TFTPTransaction
